ISSUE-054 - Merged state.d.ts and state.ts for simplicity's sake. Added new CarNotice and TyreCompound enums and data pieces.
